requirements.txt
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
nltk==3.8.1
transformers==4.38.2
torch==2.2.1
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
requests==2.31.0
geopy==2.4.1
openai>=0.27.0
yandexcloud>=0.227.0
gigachat>=0.1.0
google-generativeai>=0.3.0
dashscope>=1.10.0
deepseek-ai>=0.1.0